Add locales list
