+++ packages/strapi-plugin-i18n/admin/src/components/index.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
+export { default as LocaleRow } from './LocaleRow';
