+++ packages/strapi-plugin-i18n/admin/src/containers/SettingsPage/index.js
@@ -2,17 +2,15 @@ import React from 'react';
 import { useIntl } from 'react-intl';
 import { BaselineAlignment } from 'strapi-helper-plugin';
 import { Header, List } from '@buffetjs/custom';
-import { Pencil } from '@buffetjs/icons';
-import { Button, Text, IconLinks } from '@buffetjs/core';
-import { CustomRow } from '@buffetjs/styles';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { Button } from '@buffetjs/core';
 
+import { LocaleRow } from '../../components';
 import { useLocales } from '../../hooks';
 import { getTrad } from '../../utils';
 
-const canUpdate = true;
+// Fake permissions
 const canCreate = true;
-const canDelete = true;
+
 const LocaleSettingsPage = () => {
   const { formatMessage } = useIntl();
   const { locales, isLoading } = useLocales();
@@ -46,53 +44,22 @@ const LocaleSettingsPage = () => {
     actions,
   };
 
+  const listTitle = formatMessage(
+    {
+      id: getTrad(`Settings.locales.list.title${locales.length > 1 ? '.plural' : '.singular'}`),
+    },
+    { number: locales.length }
+  );
+
   return (
     <>
       <Header {...headerProps} />
       <BaselineAlignment top size="3px" />
       <List
-        title={formatMessage(
-          {
-            id: getTrad(
-              `Settings.locales.list.title${locales.length > 1 ? '.plural' : '.singular'}`
-            ),
-          },
-          { number: locales.length }
-        )}
+        title={listTitle}
         items={locales}
         isLoading={isLoading}
-        customRowComponent={locale => (
-          <CustomRow onClick={() => console.log('open modal')}>
-            <td>
-              <Text>{locale.code}</Text>
-            </td>
-            <td>
-              <Text fontWeight="semiBold">{locale.displayName}</Text>
-            </td>
-            <td>
-              <Text ellipsis>
-                {locale.isDefault
-                  ? formatMessage({ id: getTrad('Settings.locales.row.default-locale') })
-                  : null}
-              </Text>
-            </td>
-            <td>
-              <IconLinks
-                links={[
-                  {
-                    icon: canUpdate ? <Pencil fill="#0e1622" /> : null,
-                    onClick: () => console.log('edit'),
-                  },
-                  {
-                    icon:
-                      canDelete && !locale.isDefault ? <FontAwesomeIcon icon="trash-alt" /> : null,
-                    onClick: () => console.log('open delete modal'),
-                  },
-                ]}
-              />
-            </td>
-          </CustomRow>
-        )}
+        customRowComponent={locale => <LocaleRow locale={locale} />}
       />
     </>
   );
